bug: fixed some function dependencies
The function that determines whether a line is a lane depends on the
slope of the line being set.  I removed the call to the slope function
because it gets set by the fit_line function as well, and re-ordered
the function calls so that fit_line sets the slope before a call to
is_lane.