feat: reduce jumping of drawn lane lines
The drawn (extrapolated) lane lines jump in the video when an edge that
is considered a lane line (but is not) is the highest detected lane
edge.  This makes the green line jump off the road.  To fix this, I set
a threshold, so the lane line won’t be drawn higher than that
threshold.  The result is that the video shows a more consistent and
less jumpy lane line.