feat: draw lane lines or extrapolate full lanes
Now you can choose whether to draw full lane lines (in green), or draw
only the detected lane lines (in red).  I also refactored hough_lines
so that it does not draw the lines, but only returns the line data.
This allowed me to call a different draw function in the pipeline
(draw_lines or draw_lanes).  Draw_lines draws the detected lane lines
in red.  Draw_lanes extrapolates the full lane and draws it in green.