refactor: pipeline calls get_line_details directly; feat: pipeline intermediate images generated
Before, the function that gets line details was called by both draw
functions.  I brought it out and call it as a prior step in the
pipeline, to avoid redundant function calls in the draw_lines and
draw_lanes.

I also demonstrate the pipeline and save images of the intermediates
steps in the folder pipeline_image.